requirements.txt
gradio
youtube-transcript-api
endee
requests
python-dotenv
